Remove explicit returns in lintcss.rb
