+++ test_files/lint-bracket-missing.css
@@ -4,7 +4,7 @@
 
 .some-title {
   font-weight: lighter;
-}
+
 
 .some-title {
   font-weight: lighter;
